Use Int64 Instead Of BigInteger & No System::Numerics

+++ vs2022/Platform/Elements/07_Hyperon.h
@@ -8,17 +8,17 @@ namespace Dysnomia {
 			public ref class Hyperon : public Quark<Hadron^, Hel^, Hadron^>
 			{
 			public:
-				BigInteger Pole;
+				Int64 Pole;
 
 				Nitrogen(Hadron^ X, Hel^ C, Hadron^ Y) : Quark(C->Rod, gcnew Affinity(X, Y), C->Cone) {
 				};
 
-				BigInteger Polarize() {
+				Int64 Polarize() {
 					if (!Pole.IsZero) throw gcnew Exception("Already Polarized");
 					Pole = Math::ModPow(R->Rod->Base, R->Rod->Secret, Math::Prime);
 				}
 
-				BigInteger Depolarize() {
+				Int64 Depolarize() {
 					if (Pole.IsZero) throw gcnew Exception("Already Depolarized");
 					Pole = 0;
 				}
